@@ -3,27 +3,25 @@ name = "mlops-project"
 version = "0.1.0"
 description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 dependencies = [
-    "alembic>=1.17.2",
-    "asyncpg>=0.31.0",
+    "alembic~=1.17.0",
+    "asyncpg~=0.31.0",
     "bcrypt==4.0.1",
-    "fastapi>=0.128.0",
-    "ollama>=0.6.1",
-    "openai>=2.21.0",
-    "openpyxl>=3.1.5",
-    "passlib[bcrypt]>=1.7.4",
-    "pgvector>=0.4.2",
-    "psycopg2-binary>=2.9.11",
-    "psycopg[binary]>=3.3.2",
-    "pydantic-settings>=2.12.0",
-    "pydantic[email]>=2.12.5",
-    "python-jose[cryptography]>=3.5.0",
+    "fastapi~=0.128.0",
+    "ollama==0.6.1",
+    "openai==2.21.0",
+    "openpyxl~=3.1.0",
+    "pgvector~=0.4.2",
+    "pwdlib[argon2]~=0.2.0",
+    "pydantic-settings~=2.12.0",
+    "pydantic[email]~=2.12.0",
+    "pyjwt[crypto]~=2.10.0",
     "python-multipart>=0.0.22",
     "python-ulid>=3.1.0",
-    "redis>=7.1.0",
-    "sqlalchemy>=2.0.45",
-    "uvicorn>=0.40.0",
+    "redis~=7.1.0",
+    "sqlalchemy[asyncio]~=2.0.45",
+    "uvicorn~=0.40.0",
 ]
 
 [tool.ruff]
@@ -40,6 +38,9 @@ select = ["E", "F", "B", "I", "UP"]
 
 # 3. 指定要忽略的规则
 ignore = ["E501"] # 比如忽略“行太长”的警告，因为有些 URL 确实很长
+exclude = [
+    "alembic/versions",  # 排除生成的迁移脚本
+]
 
 [tool.ruff.format]
 # 格式化器（类似 Black）的配置
@@ -65,12 +66,14 @@ compute = [
 ]
 dev = [
     "httpx>=0.28.1",
-    "jupyter>=1.1.1",
+    "jupyterlab>=4.5.4",
     "pytest>=9.0.2",
     "pytest-asyncio>=1.3.0",
     "pytest-cov>=7.0.0",
     "pytest-mock>=3.15.1",
     "pytest-xdist>=3.8.0",
+    "ruff>=0.15.2",
+    "ty>=0.0.18",
 ]
 [tool.uvicorn]
 reload_excludes = ["postgres_data/**"]
@@ -101,7 +104,9 @@ exclude = [
     "__pycache__",
     "build",
     "dist",
-    "tests" # 如果不想对测试代码进行极端严格的类型检查，可以排除
+    "tests", # 如果不想对测试代码进行极端严格的类型检查，可以排除
+    "alembic/versions",  # 排除自动生成的脚本,
+    "alembic/archive_versions"
 ]
 
 [tool.ty.terminal]
