@@ -49,7 +49,7 @@ class AppError(Exception):
 
     status_code = 400  # 默认 400
 
-    def __init__(self, message: str, details: dict = None):
+    def __init__(self, message: str, details: dict | None = None):
         super().__init__(message)
         self.message = message
         self.details = details or {}
